CubeF4 to v1.28

+++ sources/common/libs/STM32Cube_F4/Middlewares/ST/STM32_USB_Host_Library/Core/Inc/usbh_ctlreq.h
@@ -6,13 +6,12 @@
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; Copyright (c) 2015 STMicroelectronics.
-  * All rights reserved.</center></h2>
+  * Copyright (c) 2015 STMicroelectronics.
+  * All rights reserved.
   *
-  * This software component is licensed by ST under Ultimate Liberty license
-  * SLA0044, the "License"; You may not use this file except in compliance with
-  * the License. You may obtain a copy of the License at:
-  *                      http://www.st.com/SLA0044
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
   *
   ******************************************************************************
   */
@@ -22,7 +21,7 @@
 #define __USBH_CTLREQ_H
 
 #ifdef __cplusplus
- extern "C" {
+extern "C" {
 #endif
 
 /* Includes ------------------------------------------------------------------*/
@@ -33,8 +32,8 @@
   */
 
 /** @addtogroup USBH_LIB_CORE
-* @{
-*/
+  * @{
+  */
 
 /** @defgroup USBH_CTLREQ
   * @brief This file is the
@@ -86,34 +85,34 @@ extern uint8_t USBH_CfgDesc[512];
 /** @defgroup USBH_CTLREQ_Exported_FunctionsPrototype
   * @{
   */
-USBH_StatusTypeDef USBH_CtlReq (USBH_HandleTypeDef *phost, uint8_t *buff,
-                                uint16_t length);
+USBH_StatusTypeDef USBH_CtlReq(USBH_HandleTypeDef *phost, uint8_t *buff,
+                               uint16_t length);
 
-USBH_StatusTypeDef USBH_GetDescriptor (USBH_HandleTypeDef *phost,
-                                       uint8_t  req_type, uint16_t value_idx,
-                                       uint8_t* buff, uint16_t length);
+USBH_StatusTypeDef USBH_GetDescriptor(USBH_HandleTypeDef *phost,
+                                      uint8_t  req_type, uint16_t value_idx,
+                                      uint8_t *buff, uint16_t length);
 
-USBH_StatusTypeDef USBH_Get_DevDesc (USBH_HandleTypeDef *phost, uint8_t length);
+USBH_StatusTypeDef USBH_Get_DevDesc(USBH_HandleTypeDef *phost, uint16_t length);
 
-USBH_StatusTypeDef USBH_Get_StringDesc (USBH_HandleTypeDef *phost,
-                                        uint8_t string_index, uint8_t *buff,
-                                        uint16_t length);
+USBH_StatusTypeDef USBH_Get_StringDesc(USBH_HandleTypeDef *phost,
+                                       uint8_t string_index, uint8_t *buff,
+                                       uint16_t length);
 
-USBH_StatusTypeDef USBH_SetCfg (USBH_HandleTypeDef *phost, uint16_t cfg_idx);
+USBH_StatusTypeDef USBH_SetCfg(USBH_HandleTypeDef *phost, uint16_t cfg_idx);
 
-USBH_StatusTypeDef USBH_Get_CfgDesc (USBH_HandleTypeDef *phost,uint16_t length);
+USBH_StatusTypeDef USBH_Get_CfgDesc(USBH_HandleTypeDef *phost, uint16_t length);
 
-USBH_StatusTypeDef USBH_SetAddress (USBH_HandleTypeDef *phost,
-                                    uint8_t DeviceAddress);
+USBH_StatusTypeDef USBH_SetAddress(USBH_HandleTypeDef *phost,
+                                   uint8_t DeviceAddress);
 
-USBH_StatusTypeDef USBH_SetInterface (USBH_HandleTypeDef *phost, uint8_t ep_num,
-                                      uint8_t altSetting);
+USBH_StatusTypeDef USBH_SetInterface(USBH_HandleTypeDef *phost, uint8_t ep_num,
+                                     uint8_t altSetting);
 
-USBH_StatusTypeDef USBH_SetFeature (USBH_HandleTypeDef *phost, uint8_t wValue);
+USBH_StatusTypeDef USBH_SetFeature(USBH_HandleTypeDef *phost, uint8_t wValue);
 
-USBH_StatusTypeDef USBH_ClrFeature (USBH_HandleTypeDef *phost, uint8_t ep_num);
+USBH_StatusTypeDef USBH_ClrFeature(USBH_HandleTypeDef *phost, uint8_t ep_num);
 
-USBH_DescHeader_t *USBH_GetNextDesc (uint8_t *pbuf, uint16_t *ptr);
+USBH_DescHeader_t *USBH_GetNextDesc(uint8_t *pbuf, uint16_t *ptr);
 /**
   * @}
   */
@@ -133,9 +132,7 @@ USBH_DescHeader_t *USBH_GetNextDesc (uint8_t *pbuf, uint16_t *ptr);
   */
 
 /**
-* @}
-*/
-
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
+  * @}
+  */
